requirements.txt
Django~=4.2.7
factory_boy~=3.3.0
Faker~=19.13.0
python-dotenv~=1.0.0
djangorestframework~=3.14.0
django-filter~=23.4
django-debug-toolbar~=4.2.0